서버코드 정돈 및 민경누나 요구 반영 , 스터디 상세페이지는 로그인없이 들어갈수있고, 참여는 로그인을 해야 들어갈수있음 그외. 프론트단에서 서버에서 리턴되는 특정코드를 받아 처리할게 있다 및 제주 좌표DB 구축중중
