+++ server/app.js
@@ -1,40 +1,39 @@
-const express = require("express");
-const morgan = require("morgan");
-const path = require("path");
-const dotenv = require("dotenv");
-const cors = require("cors");
-const cookieParser = require("cookie-parser");
+const express = require('express');
+const morgan = require('morgan');
+const path = require('path');
+const dotenv = require('dotenv');
+const cors = require('cors');
+const cookieParser = require('cookie-parser');
 
-dotenv.config({ path: "../.env" });
+dotenv.config({ path: '../.env' });
 //dotenv.config({ path: "../../configs/.env" });
 
-const pageRouter = require("./routes/page");
-const authRouter = require("./routes/auth");
-const userRouter = require("./routes/user");
-const apiRouter = require("./routes/api");
+const pageRouter = require('./routes/page');
+const authRouter = require('./routes/auth');
+const apiRouter = require('./routes/api');
 
-const { sequelize } = require("./models");
+const { sequelize } = require('./models');
 
 const app = express();
 
 //app.set("port", process.env.PORT || 443);
 
-app.set("port", process.env.PORT || 4000);
+app.set('port', process.env.PORT || 4000);
 
 sequelize
   .sync({ force: false })
   .then(() => {
-    console.log("데이터베이스 연결 성공");
+    console.log('데이터베이스 연결 성공');
   })
   .catch((err) => {
     console.error(err);
   });
 
-app.use(morgan("dev"));
-app.use(express.static(path.join(__dirname, "public")));
+app.use(morgan('dev'));
+app.use(express.static(path.join(__dirname, 'public')));
 //react build파일 접근경로처리
 //app.use(express.static(path.join(__dirname, "../client/build")));
-app.use("/img", express.static(path.join(__dirname, "uploads")));
+app.use('/img', express.static(path.join(__dirname, 'uploads')));
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 
@@ -54,13 +53,12 @@ app.use(cookieParser());
 app.use(cors());
 
 //swagger.js
-const { swaggerUi, specs } = require("./modules/swagger");
-app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(specs));
+const { swaggerUi, specs } = require('./modules/swagger');
+app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(specs));
 
-app.use("/", pageRouter);
-app.use("/api/auth", authRouter);
-app.use("/user", userRouter);
-app.use("/api", apiRouter);
+app.use('/', pageRouter);
+app.use('/api/auth', authRouter);
+app.use('/api', apiRouter);
 
 app.use((req, res, next) => {
   const error = new Error(`${req.method} ${req.url} 라우터가 없습니다.`);
@@ -70,13 +68,10 @@ app.use((req, res, next) => {
 
 app.use((err, req, res, next) => {
   res.locals.message = err.message;
-  res.locals.error = process.env.NODE_ENV !== "production" ? err : {};
+  res.locals.error = process.env.NODE_ENV !== 'production' ? err : {};
   res.status(err.status || 500);
 });
 
-app.listen(app.get("port"), () => {
-  console.log(
-    app.get("port"),
-    "번 포트에서 대기중...도커 젠킨스 확인 갱신확인겸"
-  );
+app.listen(app.get('port'), () => {
+  console.log(app.get('port'), '번 포트에서 대기중...도커 젠킨스 확인 갱신확인겸');
 });
