+++ server/dao/userAgent/userAgentMiddleWare.js
@@ -1,16 +1,14 @@
-const serverLoggingDB = require("../models/serverLoggingDB");
+const serverLoggingDB = require('../../models/serverLoggingDB');
 
 module.exports = function userAgentMiddleWare(serverLog) {
   return async (req, res, next) => {
     try {
       await serverLoggingDB.create({
-        serverLog: req.header("User-Agent") + serverLog,
+        serverLog: req.header('User-Agent') + serverLog,
       });
     } catch (err) {
       await serverLoggingDB.create({
-        serverLog: `error occured : ${
-          req.header("User-Agent") + serverLog
-        } err : ${err}`,
+        serverLog: `error occured : ${req.header('User-Agent') + serverLog} err : ${err}`,
       });
     }
     next();
