--- server/routes/user.js
@@ -1,8 +1,0 @@
-const express = require("express");
-
-//const { isLoggedIn } = require("./middlewares");
-const User = require("../models/user");
-
-const router = express.Router();
-
-module.exports = router;
